funcion boton quitar servicios

+++ src/java/Persistencia/ProductoJpaController.java
@@ -5,17 +5,21 @@
  */
 package Persistencia;
 
-import DTO.Producto;
-import Persistencia.exceptions.NonexistentEntityException;
-import Persistencia.exceptions.PreexistingEntityException;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import DTO.MarcaProducto;
+import DTO.DetallesProducto;
+import DTO.Producto;
+import Persistencia.exceptions.IllegalOrphanException;
+import Persistencia.exceptions.NonexistentEntityException;
+import Persistencia.exceptions.PreexistingEntityException;
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
 
 /**
  *
@@ -33,11 +37,38 @@ public class ProductoJpaController implements Serializable {
     }
 
     public void create(Producto producto) throws PreexistingEntityException, Exception {
+        if (producto.getDetallesProductoList() == null) {
+            producto.setDetallesProductoList(new ArrayList<DetallesProducto>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            MarcaProducto idMarca = producto.getIdMarca();
+            if (idMarca != null) {
+                idMarca = em.getReference(idMarca.getClass(), idMarca.getId());
+                producto.setIdMarca(idMarca);
+            }
+            List<DetallesProducto> attachedDetallesProductoList = new ArrayList<DetallesProducto>();
+            for (DetallesProducto detallesProductoListDetallesProductoToAttach : producto.getDetallesProductoList()) {
+                detallesProductoListDetallesProductoToAttach = em.getReference(detallesProductoListDetallesProductoToAttach.getClass(), detallesProductoListDetallesProductoToAttach.getId());
+                attachedDetallesProductoList.add(detallesProductoListDetallesProductoToAttach);
+            }
+            producto.setDetallesProductoList(attachedDetallesProductoList);
             em.persist(producto);
+            if (idMarca != null) {
+                idMarca.getProductoList().add(producto);
+                idMarca = em.merge(idMarca);
+            }
+            for (DetallesProducto detallesProductoListDetallesProducto : producto.getDetallesProductoList()) {
+                Producto oldIdProductoOfDetallesProductoListDetallesProducto = detallesProductoListDetallesProducto.getIdProducto();
+                detallesProductoListDetallesProducto.setIdProducto(producto);
+                detallesProductoListDetallesProducto = em.merge(detallesProductoListDetallesProducto);
+                if (oldIdProductoOfDetallesProductoListDetallesProducto != null) {
+                    oldIdProductoOfDetallesProductoListDetallesProducto.getDetallesProductoList().remove(detallesProductoListDetallesProducto);
+                    oldIdProductoOfDetallesProductoListDetallesProducto = em.merge(oldIdProductoOfDetallesProductoListDetallesProducto);
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             if (findProducto(producto.getCodigo()) != null) {
@@ -51,12 +82,33 @@ public class ProductoJpaController implements Serializable {
         }
     }
 
-    public void edit(Producto producto) throws NonexistentEntityException, Exception {
+    public void edit(Producto producto) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Producto persistentProducto = em.find(Producto.class, producto.getCodigo());
+            MarcaProducto idMarcaOld = persistentProducto.getIdMarca();
+            MarcaProducto idMarcaNew = producto.getIdMarca();
+            List<String> illegalOrphanMessages = null;
+            
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            if (idMarcaNew != null) {
+                idMarcaNew = em.getReference(idMarcaNew.getClass(), idMarcaNew.getId());
+                producto.setIdMarca(idMarcaNew);
+            }
             producto = em.merge(producto);
+            if (idMarcaOld != null && !idMarcaOld.equals(idMarcaNew)) {
+                idMarcaOld.getProductoList().remove(producto);
+                idMarcaOld = em.merge(idMarcaOld);
+            }
+            if (idMarcaNew != null && !idMarcaNew.equals(idMarcaOld)) {
+                idMarcaNew.getProductoList().add(producto);
+                idMarcaNew = em.merge(idMarcaNew);
+            }
+            
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -74,7 +126,7 @@ public class ProductoJpaController implements Serializable {
         }
     }
 
-    public void destroy(String id) throws NonexistentEntityException {
+    public void destroy(String id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -85,6 +137,22 @@ public class ProductoJpaController implements Serializable {
                 producto.getCodigo();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The producto with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<DetallesProducto> detallesProductoListOrphanCheck = producto.getDetallesProductoList();
+            for (DetallesProducto detallesProductoListOrphanCheckDetallesProducto : detallesProductoListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Producto (" + producto + ") cannot be destroyed since the DetallesProducto " + detallesProductoListOrphanCheckDetallesProducto + " in its detallesProductoList field has a non-nullable idProducto field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            MarcaProducto idMarca = producto.getIdMarca();
+            if (idMarca != null) {
+                idMarca.getProductoList().remove(producto);
+                idMarca = em.merge(idMarca);
             }
             em.remove(producto);
             em.getTransaction().commit();
